Structuring tests.
